+++ app-web/src/interfaces/weatherForecast.interface.ts
@@ -6,6 +6,7 @@ export interface IWeatherForecast {
     humidity?: number,
     pressure?: number,
     tomorowTemperature?: ITemperature
+    weatherIcon?: string,
     afterTomorrowTemperature?: ITemperature
 }